+++ tests/test_metrics.py
@@ -1,15 +1,12 @@
 from pathlib import Path
 
-import matplotlib
+import numpy
 import pandas
+import pytest
+from matplotlib import pyplot
 from numpy.random.mtrand import RandomState
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.tree import DecisionTreeClassifier
-
-matplotlib.use('agg')
-import numpy
-import pytest
-from matplotlib import pyplot
 
 from ds_utils.metrics import plot_confusion_matrix, plot_metric_growth_per_labeled_instances, \
     visualize_accuracy_grouped_by_probability
@@ -22,6 +19,9 @@ x_test = pandas.read_csv(Path(__file__).parents[0].absolute().joinpath("resource
 y_train = pandas.read_csv(Path(__file__).parents[0].absolute().joinpath("resources").joinpath("iris_y_train.csv"))
 y_test = pandas.read_csv(Path(__file__).parents[0].absolute().joinpath("resources").joinpath("iris_y_test.csv"))
 
+Path(__file__).parents[0].absolute().joinpath("result_images").mkdir(exist_ok=True)
+Path(__file__).parents[0].absolute().joinpath("result_images").joinpath("test_metrics").mkdir(exist_ok=True)
+
 
 def test_print_confusion_matrix_binary():
     custom_y_test = "1 1 1 1 1 0 1 0 1 0 0 1 0 0 1 0 0 0 1 0 1 0 0 1 1 1 1 1 0 0 0 1 0 1 0 1 0 0 0 0 1 1 1 1 0 1 1 " \
@@ -33,8 +33,6 @@ def test_print_confusion_matrix_binary():
                           numpy.fromstring(custom_y_pred, dtype=int, sep=' '),
                           [1, 0])
 
-    Path(__file__).parents[0].absolute().joinpath("result_images").mkdir(exist_ok=True)
-    Path(__file__).parents[0].absolute().joinpath("result_images").joinpath("test_metrics").mkdir(exist_ok=True)
     result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath("test_metrics").joinpath(
         "test_print_confusion_matrix_binary.png")
     pyplot.savefig(str(result_path))
@@ -55,8 +53,7 @@ def test_print_confusion_matrix():
     plot_confusion_matrix(numpy.fromstring(custom_y_test, dtype=int, sep=' '),
                           numpy.fromstring(y_pred, dtype=int, sep=' '),
                           [0, 1, 2])
-    Path(__file__).parents[0].absolute().joinpath("result_images").mkdir(exist_ok=True)
-    Path(__file__).parents[0].absolute().joinpath("result_images").joinpath("test_metrics").mkdir(exist_ok=True)
+
     result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath("test_metrics").joinpath(
         "test_print_confusion_matrix.png")
     pyplot.savefig(str(result_path))