+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.7] - 2020-09-14
+### Added
+- xai::plot_features_importance method that visualize into bar chart the feature importance.
+- a new module named `unsupervised` was added. The module contains methods that calculate and/or visualize evaluation 
+performance of an unsupervised model.
+- unsupervised::plot_cluster_cardinality method that plots the number of points per cluster as a bar chart.
+- unsupervised::plot_cluster_magnitude method that plots the Total Point-to-Centroid Distance per cluster as a bar 
+chart.
+- unsupervised::plot_magnitude_vs_cardinality method plots the cardinality vs. magnitude as a scatter plot.
+- unsupervised::plot_loss_vs_cluster_number method that plots the graph which helps to find the optimum parameter ``k``
+for KMeans.
+### Changed
+- deprecated xai::draw_tree. Use sklearn.tree.plot_tree instead.
+- requirements dependencies.
+### Fixed
+- minor changes
+
 ## [1.6.3] - 2020-07-09
 ### Added
 - code examples to README.md